Add files via upload
This is a variation of the Fibonacci series where the first three terms—2, 2, and 3—remain constant, while each subsequent term is the sum of the previous two terms. For example, 2 + 2 = 4, so the fourth term is 4. Similarly, the fifth term is 2 + 3 = 5, and the sixth term is 3 + 4 = 7. Thus, the Varionacci series for four terms is 2, 2, 3, 4; for five terms, it is 2, 2, 3, 4, 5; for six terms, it is 2, 2, 3, 4, 5, 7; and so on.